chore: rename cmsConfig file for consistancy
